+++ Classes/DependencyInjection/Widgets/CountriesRegistration.php
@@ -102,6 +102,7 @@ final class CountriesRegistration extends AbstractRegistration
             ->arg(
                 '$options',
                 [
+                    'reportLink' => $this->buildReportLink(),
                     'siteTitle' => $this->matomoConfiguration->getSiteTitle(),
                     'title' => $localisedTitle,
                 ]
@@ -118,5 +119,14 @@ final class CountriesRegistration extends AbstractRegistration
                     'width' => 'small',
                 ]
             );
+    }
+
+    private function buildReportLink(): string
+    {
+        return \sprintf(
+            '%s?module=CoreHome&action=index&idSite=%d&period=month&date=today#?segment=&category=General_Visitors&subcategory=UserCountry_SubmenuLocations',
+            $this->matomoConfiguration->getUrl(),
+            $this->matomoConfiguration->getIdSite()
+        );
     }
 }