+++ Classes/DependencyInjection/Widgets/VisitsPerDayRegistration.php
@@ -76,6 +76,7 @@ final class VisitsPerDayRegistration extends AbstractRegistration
             ->arg(
                 '$options',
                 [
+                    'reportLink' => $this->buildReportLink(),
                     'siteTitle' => $this->matomoConfiguration->getSiteTitle(),
                     'title' => $localisedTitle,
                 ]
@@ -92,5 +93,14 @@ final class VisitsPerDayRegistration extends AbstractRegistration
                     'width' => 'medium',
                 ]
             );
+    }
+
+    private function buildReportLink(): string
+    {
+        return \sprintf(
+            '%s?module=CoreHome&action=index&idSite=%d&period=day&date=today#?segment=&category=General_Visitors&subcategory=General_Overview',
+            $this->matomoConfiguration->getUrl(),
+            $this->matomoConfiguration->getIdSite()
+        );
     }
 }